+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 11. Added typeDefs and resolvers for Filters, Instruments, Locations and Users
 12. Added seeds data for Users, Locations, Instruments and Filters
 13. Added Components for Header, Footer, Weather Data, Images and Filet, Instrument and Location Forms
+14. Added utils for auth as welll as mutastions and queries